# api/src/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: Bon-API
  version: v0
servers:
  -
    url: 'http://localhost:8000/api/v0'
paths:
...
  /numeroTable:
    get:
      tags:
        - Table
      operationId: 3fa68adb7e048e63928eb838bb61d69b
      responses:
        '200':
          description: 'Retourne l ensemble des tables'
          content:
            application/json:
              schema:
                description: 'Tables du restaurant'
                type: array
                items:
                  $ref: '#/components/schemas/NumeroTable'
    post:
      tags:
        - Table
      operationId: 80f050cecbf3df0a4810404254096196
      requestBody:
        content:
          application/json:
            schema:
              required:
                - nombre_de_personnes
              properties:
                nombre_de_personnes:
                  type: integer
              type: object
              example:
                nombre_de_personnes: 3
      responses:
        '200':
          description: 'Table créée'
          content:
            application/json:
              schema:
                type: string
      security:
        -
          Authorization: []
  '/numeroTable/{id}':
    get:
      tags:
        - Table
      operationId: 0097b61a3329b225c0fb2a08e3371ebc
      parameters:
        -
          name: id
          in: path
          description: 'ID de la table'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Retourne une table en fonction de son ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NumeroTable'
    put:
      tags:
        - Table
      operationId: 6d315a265d7c23fbf2b2ae00b4d1782b
      parameters:
        -
          name: id
          in: path
          description: 'ID de la table à modifier'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              required:
                - nombre_de_personnes
              properties:
                nombre_de_personnes:
                  type: integer
              type: object
              example:
                nombre_de_personnes: 3
      responses:
        '200':
          description: 'Table modifiée'
          content:
            application/json:
              schema:
                type: string
      security:
        -
          Authorization: []
    delete:
      tags:
        - Table
      operationId: 2e458eceb681c8ffabbaab634c4e2919
      parameters:
        -
          name: id
          in: path
          description: 'ID de la table'
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 'Suppression réussie'
          content:
            application/json:
              schema:
                type: string
        '404':
          description: 'Erreur de récupération'
          content:
            application/json:
              schema:
                description: 'Message d erreur'
                type: string
              example: 'Cette table n''existe pas/plus'
      security:
        -
          Authorization: []
  /platCommande:
    get:
      tags:
        - 'Plat Commande'
      operationId: da6326872318a9e5f822e855f2b3bb09
      responses:
        '200':
          description: 'Retourne l ensemble des plats commandés'
          content:
            application/json:
              schema:
                description: 'Tous les plats commandés'
                type: array
                items:
                  $ref: '#/components/schemas/PlatCommande'
    post:
      tags:
        - 'Plat Commande'
      operationId: 0b4885f2df48d26606372d6fc61209fe
      requestBody:
        content:
          application/json:
            schema:
              required:
                - plat_id
                - commande_id
              properties:
                plat_id:
                  type: integer
                commande_id:
                  type: integer
              type: object
              example:
                plat_id: 18
                commande_id: 3
      responses:
        '200':
          description: 'Plat de la commande créé'
          content:
            application/json:
              schema:
                type: string
      security:
        -
          Authorization: []
  '/platCommande/{id}':
    get:
      tags:
        - 'Plat Commande'
      operationId: a2172a39b6ca86e696cc46f21974fb92
      parameters:
        -
          name: id
          in: path
          description: 'ID du plat de la commande'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Retourne un plat d\''une commande en fonction de son ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatCommande'
    put:
      tags:
        - 'Plat Commande'
      operationId: e74080b036e3d5a36389b6c6033cad2e
      parameters:
        -
          name: id
          in: path
          description: 'ID du plat de la commande à modifier'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              required:
                - plat_id
                - commande_id
              properties:
                plat_id:
                  type: integer
                commande_id:
                  type: integer
              type: object
              example:
                plat_id: 18
                commande_id: 3
      responses:
        '200':
          description: 'Plat de la commande modifié'
          content:
            application/json:
              schema:
                type: string
      security:
        -
          Authorization: []
    delete:
      tags:
        - 'Plat Commande'
      operationId: 840d356c8f366a13d931dca51374434a
      parameters:
        -
          name: id
          in: path
          description: 'ID du plat dans la commande'
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 'Suppression réussie'
          content:
            application/json:
              schema:
                type: string
        '404':
          description: 'Erreur de récupération'
          content:
            application/json:
              schema:
                description: 'Message d erreur'
                type: string
              example: 'Ce plat de la commande n''existe pas/plus'
      security:
        -
          Authorization: []
...
components:
  schemas:
...
    NumeroTable:
      properties:
        id:
          type: integer
          nullable: false
        nombre_de_personnes:
          type: integer
          nullable: false
      type: object
...
    PlatCommande:
      properties:
        id:
          type: integer
          nullable: false
        plat_id:
          type: integer
          nullable: false
        commande_id:
          type: integer
          nullable: false
      type: object
...
  securitySchemes:
    Authorization:
      type: http
      bearerFormat: JWT
      scheme: bearer
...